Used the helper function
